Record/report state of client script/BBA for diag
Keep track of the messages, the form fields, and screenshot the browser
after each step to later allow the user to diagnose the authentication
steps through the authentication report.
Change automation to allow to enable diagnostics for client script/BBA.
Change database to expose Flyway to allow other add-ons to use it.
Change reports to allow to clean up report data and read HTTP message.

Signed-off-by: thc202 <thc202@gmail.com>